+++ tests/_test_manure_module/test_animal_manure_excretions/_test_dry_cow_manure_excretion.py
@@ -8,6 +8,7 @@ from RUFAS.biophysical.animal.data_types.animal_manure_excretions import AnimalM
 from RUFAS.general_constants import GeneralConstants
 from RUFAS.routines.animal.animal_module_constants import AnimalModuleConstants
 from RUFAS.routines.animal.manure.dry_cow_manure_excretion import manure_calculations
+from RUFAS.user_constants import UserConstants
 
 
 @pytest.mark.skip(reason="Skipping this test as AnimalManureExcretions is modified")
@@ -71,14 +72,14 @@ def test_dry_cow_manure_calculations(methane_model: str, mocker: MockerFixture) 
         15.1
         + 0.83
         * (dry_matter_intake * GeneralConstants.KG_TO_GRAMS)
-        * (CP_concentration * GeneralConstants.PROTEIN_TO_NITROGEN)
+        * (CP_concentration * UserConstants.PROTEIN_TO_NITROGEN)
         / GeneralConstants.FRACTION_TO_PERCENTAGE
     ) * GeneralConstants.GRAMS_TO_KG
     fecal_nitrogen = (
         0.345
         + 0.317
         * (dry_matter_intake * GeneralConstants.KG_TO_GRAMS)
-        * (CP_concentration * GeneralConstants.PROTEIN_TO_NITROGEN)
+        * (CP_concentration * UserConstants.PROTEIN_TO_NITROGEN)
         / GeneralConstants.FRACTION_TO_PERCENTAGE
     ) * GeneralConstants.GRAMS_TO_KG
     urine_nitrogen = manure_nitrogen - fecal_nitrogen